BridgeJS: Merge small core files into single Misc.swift
Ugry hack to allow me to add new files later without paying build
time penalty...
