--- Plugins/BridgeJS/Sources/BridgeJSCore/BridgeJSCoreError.swift
@@ -1,7 +1,0 @@
-public struct BridgeJSCoreError: Swift.Error, CustomStringConvertible {
-    public let description: String
-
-    public init(_ message: String) {
-        self.description = message
-    }
-}
